Adding Jet pack and Fix Key

+++ ProjectGame/Project Game/Assets/Scripts/Active Scripts/Scripts Luis/Key.cs
@@ -16,9 +16,11 @@ public class Key : MonoBehaviour
 
     private void OnTriggerEnter(Collider other)
     {
-        GameManager.instance.updateGameGoal(-1);
-        Objective.instance.removeDir(transform.position);
-        aud.PlayOneShot(keySound, keyVol);
-        Destroy(gameObject);
+        if (other.CompareTag("Player")) {
+            GameManager.instance.updateGameGoal(-1);
+            Objective.instance.removeDir(transform.position);
+            aud.PlayOneShot(keySound, keyVol);
+            Destroy(gameObject);
+        }
     }
 }
